% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_gather.R
\name{time_gather}
\alias{time_gather}
\title{Illustration of time_gather}
\usage{
time_gather(.data, by)
}
\arguments{
\item{.data}{a tibble containing d_min and d_max date columns}

\item{by}{a string containing the timestep, "year", "month", "day"}
}
\value{
a long data frame with a dates column
}
\description{
gather time dimension
}
