% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spread.R
\name{time_spread}
\alias{time_spread}
\title{Illustration of time_spread}
\usage{
time_spread(.data)
}
\arguments{
\item{.data}{a tibble containing an id and dates}
}
\value{
a data frame with min and max date for each id
}
\description{
gather time dimension
}
